compose/siteapp/pyproject.toml
[project]
name = "siteapp"
version = "0.1.0"
description = "Public docs + agent downloads + admin for lab-bridge."
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115,<0.116",
    "uvicorn[standard]>=0.30,<0.31",
    "jinja2>=3.1,<4",
    "markdown-it-py[linkify]>=3.0,<4",
    "mdit-py-plugins>=0.4,<0.5",
    "pygments>=2.18,<3",
    "python-multipart>=0.0.9,<0.1",
    "itsdangerous>=2.2,<3",
]

[dependency-groups]
dev = [
    "pytest>=8.3,<9",
    "httpx>=0.27,<0.29",
    "pytest-asyncio>=0.24,<0.25",
    "ruff>=0.6,<0.13",
]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py313"
